feat: Add edge deletion and update API for UpdateTransaction (#1169)
#1138 
Abandon the previous store changes in UpdateTransaction manner, modify
the changes in place, and add implementation for UndoLog.

This pull request makes significant changes to the graph storage and
transaction interfaces, focusing on improving edge property management
and simplifying edge iterator APIs. The most important changes include
replacing the legacy edge iterator and edge data update methods with new
offset-based APIs, updating related documentation, and refactoring undo
log types for consistency. These changes streamline edge property
updates and deletions, making the interfaces more robust and
maintainable.

### API and Interface Refactoring

* Removed legacy `SetEdgeData`, `GetUpdatedEdgeData`,
`GetOutEdgeIterator`, and `GetInEdgeIterator` methods from
`StorageUpdateInterface` and `UpdateTransaction`, replacing them with
new offset-based methods such as `UpdateEdgeProperty`, `DeleteEdge`, and
simplified graph view accessors.
[[1]](diffhunk://#diff-f87627fb15908d9a809488cbe5e18343f831fd13dab951809413c0cbd01dc8e6L234-R240)
[[2]](diffhunk://#diff-f87627fb15908d9a809488cbe5e18343f831fd13dab951809413c0cbd01dc8e6L265-R274)
[[3]](diffhunk://#diff-a68bd0cc716f740f4017577bb3a904d6ad19fed7c839f475463c6770eb43f307L234-R266)
* Refactored undo log structures (`InsertEdgeUndo`,
`UpdateEdgePropUndo`, `RemoveEdgeUndo`) to use `int32_t` for offsets
instead of `size_t` for consistency and correctness.
[[1]](diffhunk://#diff-74e0f62ecb5d31111e55bc8799ee11da5045ac2b0f67fd61f94b7d3377f9c6e4L66-R69)
[[2]](diffhunk://#diff-74e0f62ecb5d31111e55bc8799ee11da5045ac2b0f67fd61f94b7d3377f9c6e4L96-R101)
[[3]](diffhunk://#diff-74e0f62ecb5d31111e55bc8799ee11da5045ac2b0f67fd61f94b7d3377f9c6e4L126-R129)

### Edge Property and Deletion Enhancements

* Added new methods to `PropertyGraph` and `UpdateTransaction` for
deleting edges and updating edge properties using explicit offsets,
improving control and reliability for edge operations.
[[1]](diffhunk://#diff-ca0b2344754e05e932c780324554dfcfd96455ccd98516af2fb5913023282eeeR245-R248)
[[2]](diffhunk://#diff-ca0b2344754e05e932c780324554dfcfd96455ccd98516af2fb5913023282eeeR314-R319)
[[3]](diffhunk://#diff-a68bd0cc716f740f4017577bb3a904d6ad19fed7c839f475463c6770eb43f307R225-R230)
* Updated operation type enums to clarify that edge property updates and
removals are now offset-based.

### Documentation Updates

* Removed documentation for deprecated iterator and edge data methods,
and added documentation for new offset-based edge property update
methods in `doc/source/reference/cpp_api/gs.md`.
[[1]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL1145-L1160)
[[2]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL1412-L1427)
[[3]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL1443-L1490)
[[4]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL21856-L21857)
[[5]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL21874-R21778)
* Removed documentation for deprecated `edge_iterator` classes in both
read and update transaction sections.
[[1]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL3493-L3509)
[[2]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL4382-L4401)
[[3]](diffhunk://#diff-da5de363d81bf4eef427f8424064d2afd3e3b6fc085ceddd6011d8601df16bbfL24500-L24517)

### Utility and Supporting Changes

* Added new utility functions for searching edge offsets in
`generic_view_utils.h`, supporting the new offset-based APIs.
* Added methods to `PropertyGraph` for accessing edge tables by label
triplet, supporting the new API structure.

These changes collectively modernize and simplify the edge property
management workflow, reduce API surface complexity, and improve
documentation consistency.

+++ include/neug/transaction/undo_log.h
@@ -63,10 +63,10 @@ struct InsertVertexUndo : public IUndoLog {
 struct InsertEdgeUndo : public IUndoLog {
   label_t src_label, dst_label, edge_label;
   vid_t src_lid, dst_lid;
-  size_t oe_offset, ie_offset;
+  int32_t oe_offset, ie_offset;
   InsertEdgeUndo(label_t src_label, label_t dst_label, label_t edge_label,
-                 vid_t src_lid, vid_t dst_lid, size_t oe_offset,
-                 size_t ie_offset)
+                 vid_t src_lid, vid_t dst_lid, int32_t oe_offset,
+                 int32_t ie_offset)
       : src_label(src_label),
         dst_label(dst_label),
         edge_label(edge_label),
@@ -93,12 +93,12 @@ struct UpdateVertexPropUndo : public IUndoLog {
 struct UpdateEdgePropUndo : public IUndoLog {
   label_t src_label, dst_label, edge_label;
   vid_t src_lid, dst_lid;
-  size_t oe_offset, ie_offset;
+  int32_t oe_offset, ie_offset;
   int32_t col_id;
   Property value;
   UpdateEdgePropUndo(label_t src_label, label_t dst_label, label_t edge_label,
-                     vid_t src_lid, vid_t dst_lid, size_t oe_offset,
-                     size_t ie_offset, int32_t col_id, const Property& value)
+                     vid_t src_lid, vid_t dst_lid, int32_t oe_offset,
+                     int32_t ie_offset, int32_t col_id, const Property& value)
       : src_label(src_label),
         dst_label(dst_label),
         edge_label(edge_label),
@@ -123,10 +123,10 @@ struct RemoveVertexUndo : public IUndoLog {
 struct RemoveEdgeUndo : public IUndoLog {
   label_t src_label, dst_label, edge_label;
   vid_t src_lid, dst_lid;
-  size_t oe_offset, ie_offset;
+  int32_t oe_offset, ie_offset;
   RemoveEdgeUndo(label_t src_label, label_t dst_label, label_t edge_label,
-                 vid_t src_lid, vid_t dst_lid, size_t oe_offset,
-                 size_t ie_offset)
+                 vid_t src_lid, vid_t dst_lid, int32_t oe_offset,
+                 int32_t ie_offset)
       : src_label(src_label),
         dst_label(dst_label),
         edge_label(edge_label),
